+++ src/index.js
@@ -1,7 +1,7 @@
 import Localization from './Localization';
-import LocalizationConnector from './LocalizationConnector';
+import localizationConnector from './LocalizationConnector';
 import LocalizationWrapper from './LocalizationWrapper';
 import Text from './Text';
 
 export default Localization;
-export { Text, LocalizationWrapper, LocalizationConnector };
+export { Text, LocalizationWrapper, localizationConnector };